Move index redirects into InstallRedirect
