+++ components/form/ResourceTabs/index.vue
@@ -39,7 +39,7 @@ export default {
     const inStore = this.$store.getters['currentProduct'].inStore;
 
     if ( this.$store.getters[`${ inStore }/schemaFor`](EVENT) ) {
-      this.hasEvents = true;
+      this.hasEvents = true; // @TODO be smarter about which ones actually ever have events
       this.allEvents = await this.$store.dispatch(`${ inStore }/findAll`, { type: EVENT });
     }
   },
@@ -112,11 +112,11 @@ export default {
   <Tabbed v-if="!isView || hasCustomTabs || hasConditions" v-bind="$attrs">
     <slot />
 
-    <Tab v-if="hasConditions" :label="t('resourceTabs.tabs.conditions')" name="conditions" :weight="11">
+    <Tab v-if="hasConditions" :label="t('resourceTabs.tabs.conditions')" name="conditions">
       <Conditions :value="value" />
     </Tab>
 
-    <Tab v-if="!$fetchState.pending && hasEvents" :label="t('resourceTabs.tabs.events')" name="events" :weight="10">
+    <Tab v-if="!$fetchState.pending && hasEvents" :label="t('resourceTabs.tabs.events')" name="events">
       <SortableTable
         :rows="events"
         :headers="eventHeaders"
